requirements.txt
absl-py==2.0.0
albumentations==1.3.1
antlr4-python3-runtime==4.8
astunparse==1.6.3
audioread==3.0.1
av==11.0.0
beautifulsoup4==4.12.3
bitarray==2.8.2
boto3==1.28.78
botocore==1.31.78
cachetools==4.2.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.30.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.5
decorator==4.4.2
decord==0.6.0
einops==0.7.0
einops-exts==0.0.4
fairseq==0.12.2
filelock==3.13.1
flatbuffers==23.5.26
flow-vis==0.1
fonttools==4.44.0
fsspec==2023.10.0
gast==0.4.0
gdown==5.1.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
huggingface-hub==0.17.3
hydra-core==1.0.7
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
json-tricks==3.17.3
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.7.1
lit==18.1.8
llvmlite==0.41.0
lpips==0.1.4
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
natsort==8.4.0
networkx==3.1
numba==0.58.0
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.0.5
opencv-contrib-python==4.8.0.76
opencv-python==4.7.0.72
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
portalocker==2.8.2
proglog==0.1.10
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.1.1
PySocks==1.7.1
pyspng==0.1.1
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rotary-embedding-torch==0.3.5
rsa==4.9
s3transfer==0.7.0
sacrebleu==2.3.1
sacremoses==0.1.1
safetensors==0.4.0
scenedetect==0.5.1
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.9.1
sentencepiece==0.1.99
six==1.16.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
sympy==1.13.1
sync-batchnorm==0.0.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.14.1
torch==2.1.0+cu121
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.18
visualize==0.5.1
Werkzeug==3.0.1
wrapt==1.15.0
zipp==3.17.0
